Tathmini_Backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
email-validator>=2.0.0
motor>=3.3.1
pymongo>=4.6.0
httpx>=0.25.0
cryptography>=41.0.4